service/pyproject.toml
[project]
name = "lumicore"
version = "0.1.0"
description = ""
authors = [
    {name = "Anderson Freitas",email = "freitas.dev@proton.me"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "materialdash (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-simple-history (>=3.8.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"